Create README.md
Database Connection with Singleton

I use Singleton to connect Database, because Singleton is only applicable in database connection in this project (is the only one connection that we use in this assignment).  The other reason is its more convenient..

Singleton pattern ensure only one instance of a class is instantiated for the lifetime of the application.

Singleton pattern is not much recommended Design pattern, because its globally accessible and have many drawbacks.  


Builder Design Pattern

Builder Design Pattern provides clear separation between construction and representation of an object. It provides better control over construction process and supports to change the internal representation of objects

The Builder Design Pattern helps us to slice the operations of building an object. It focuses on constructing a complex object step by step. It also enforces a process to create an object as a finished product. That means an object has to be massaged by some instructed steps before it is ready and can be used by others.